requirements.txt
# langchain
# langchain_community
# langchain_experimental
# openai
# pandas
# openpyxl
# python-dotenv
# tabulate


langchain 
langchain-openai 
langchain-experimental 
pandas 
python-dotenv 
openpyxl
psycopg2-binary
